Fix the bug in `sigchld_handler()`
SIGCHLD is delivered by the kernel to the parent (shell) process
when any of its children has terminated or was stopped by a signal.
In our fixes, we add an extra `WNOHANG` option in `waitpid()` which
allows the handler not to block. Without it, a deadlock might occur.

For example, run 'cat | tr a-z A-Z | sort' in the shell, enter some
text and then enter Ctrl-Z, the SIGTSTP signal will be delivered to
each of these three processes. First SIGTSTP signal arrives and then
forces `sigchld_handler()` to run in the parent. In the while loop
`waitpid()` will be executed three times, some time later, the 2nd
or the pending SIGTSTP signal will force the `sigchld_handler()`
to be run again. But these three child processes have all been
`waitpid()`ed, thus this call to `waitpid()` will block forever!

So, we shouldn't block in `waitpid()` (by adding `WNOHANG` option).
In previous implementation, we break the while loop each time the
`sigchld_handler()` is run because of `SIGTSTP`, i.e., the shell
just waited one `SIGTSTP` then exits the handler. But note that
pending signals are not queued (see 'CSAPP_3e') and therefore the
3rd SIGCHLD signal may be discarded, which accounts for why Ctrl-Z
sometimes is broken (see previous commit:
https://github.com/How-u-doing/msh/commit/ba85eb4).

Lastly, `sigchld_handler()` will not run when SIGCONT is delivered.
Hence, there's no need to handle it in the handler.