Fix `cmd 2>&1 >out | do_sth_with_stderr` problem
Previously, only `>out 2>&1` was implemented, the reverse
with the order exchanged was treated as the same. However,
`2>&1 >out` can be quite helpful if we want to pipe only
stderr, for example,

    command 2>&1 >/dev/null | grep "something"

This first sets up the pipe, then redirects stderr to where
stdout (fd 1) is referring to (i.e., the write end of the pipe)
and then redirects stdout to /dev/null (do not output stdout),
which ensures only stderr is piped.