Initialize sigset before sigadd() and fix typos
Since `sigset_t` is a struct comprised of an array of primitive/POD
objects (unsigned long int), in statement `sigset_t mask;` mask
won't get initialized. Thus, we need to do it manually by calling
sigemptyset() to initialize the set to the empty set or calling
sigsetfill() to add every signal to the set. Of course, we can
also perform the zero-initialization in the C++ish way by using
an empty brace to zero initialize the set, i.e.
`sigset_t mask{};` or `sigset_t mask = {};`.

Given the existence of `sigemptyset()`, I'd prefer it (as the
traditional way to initialize sigset_t objects).